Fix out-of-tree builds
There were two places where the mvapich2 build seemed to break
when doing out of tree builds, e.g.

 mkdir BUILD; cd BUILD; ../configure && make

In one case, the f90 bindings failed to build because -I${srcdir} was
not in FCFLAGS. In the other case, osu_benchmarks failed to build because
the MPI library couldn't be found. This latter case was fixed by appending
../lib to LD_LIBRARY_PATH.